requirements.txt
mysql-connector-python
sqlalchemy
ipaddress
pymysql
tomli
